# docs/swagger.yaml
basePath: /
definitions:
...
  main.discordMemberRow:
    properties:
      avatar:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Admin panel API for a Dune Awakening private server.
    When auth_enabled is set, all /api/v1 endpoints (except /api/v1/auth/*) require the session cookie issued by /api/v1/auth/login or the Discord OAuth flow.
  title: dune-admin API
  version: "1.0"
paths:
...
  /api/v1/discord/members/search:
    get:
      parameters:
      - description: Name prefix to search
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.discordMemberRow'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search Discord guild members by name prefix
      tags:
      - discord
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: dune_admin_session
    type: apiKey
swagger: "2.0"